(svn r25340) [1.3] -Backport from trunk: - Fix: Loading only 8 bits into a 16 bit variable could cause endianness problems (r25337) - Fix: Check for zero width space in translations and fail upon finding them [FS#5589] (r25326) - Fix: [SDL] Keyboard input stopped working after fullscreen toggle [FS#5580] (r25318) - Fix: Proper size-estimation for numbers with n digits, i.e. not assume a particular number is the widest [FS#5562] (r25314, r25313)

+++ src/saveload/station_sl.cpp
@@ -226,7 +226,7 @@ static const SaveLoad _old_station_desc[] = {
 static uint16 _waiting_acceptance;
 static uint16 _cargo_source;
 static uint32 _cargo_source_xy;
-static uint16 _cargo_days;
+static uint8  _cargo_days;
 static Money  _cargo_feeder_share;
 
 static const SaveLoad _station_speclist_desc[] = {
